requirements.txt
pydantic~=1.8.2
asyncpg==0.22.0
fastapi==0.63.0
psycopg2-binary==2.8.6
uvicorn==0.13.4
fastapi-utils==0.2.1
pytz